BYT-184: Remove redundancy
