+++ backend/app/src/controllers/yelp/comments.ts
@@ -26,7 +26,7 @@ export const getComments: RequestHandler = async (req, res, next) => {
       !req.query.endDate ||
       req.query.endDate.length !== 8
     )
-      return res.status(400).send();
+      return res.status(400).send(invalidInput);
 
     const user = await User.findOne({
       where: { username: req.session.username },
