+++ frontend/app/src/Components/Tables/CommentsTable/CommentsTableColDefConst.tsx
@@ -184,11 +184,101 @@ const twitterColDef: MetricsTableColDef = [
   },
 ];
 
+const googleReviewsColDef: MetricsTableColDef = [
+  {
+    field: 'reviewer',
+    headerName: 'Name',
+    ...defaultColumnStyle,
+    ...defaultStringFilterStyle,
+  },
+  {
+    field: 'title',
+    headerName: 'Title',
+    ...defaultColumnStyle,
+    ...defaultStringFilterStyle,
+  },
+  {
+    field: 'review',
+    headerName: 'Review',
+    flex: 1,
+    ...defaultColumnStyle,
+    ...defaultStringFilterStyle,
+  },
+  {
+    field: 'rating',
+    headerName: 'Rating',
+    type: 'number',
+    ...defaultColumnStyle,
+    ...defaultNumberFilterStyle,
+  },
+  {
+    field: 'sentimentAnalysis',
+    headerName: 'Sentiment',
+    ...defaultColumnStyle,
+    ...defaultStringFilterStyle,
+  },
+  {
+    field: 'subjectivityAnalysis',
+    headerName: 'Subjectivity',
+    ...defaultColumnStyle,
+    ...defaultStringFilterStyle,
+  },
+  {
+    field: 'topicClassification',
+    headerName: 'Topic',
+    ...defaultColumnStyle,
+    ...defaultStringFilterStyle,
+  },
+];
+
+const yelpColDef: MetricsTableColDef = [
+  {
+    field: 'userName',
+    headerName: 'Name',
+    ...defaultColumnStyle,
+    ...defaultStringFilterStyle,
+  },
+  {
+    field: 'text',
+    headerName: 'Review',
+    flex: 1,
+    ...defaultColumnStyle,
+    ...defaultStringFilterStyle,
+  },
+  {
+    field: 'rating',
+    headerName: 'Rating',
+    type: 'number',
+    ...defaultColumnStyle,
+    ...defaultNumberFilterStyle,
+  },
+  {
+    field: 'sentimentAnalysis',
+    headerName: 'Sentiment',
+    ...defaultColumnStyle,
+    ...defaultStringFilterStyle,
+  },
+  {
+    field: 'subjectivityAnalysis',
+    headerName: 'Subjectivity',
+    ...defaultColumnStyle,
+    ...defaultStringFilterStyle,
+  },
+  {
+    field: 'topicClassification',
+    headerName: 'Topic',
+    ...defaultColumnStyle,
+    ...defaultStringFilterStyle,
+  },
+];
+
 export const CommentsTableColDefConst = {
   Facebook: defaultColDef,
   Instagram: instagramColDef,
   Twitter: twitterColDef,
   YouTube: defaultColDef,
   Reddit: redditColDef,
+  GoogleReviews: googleReviewsColDef,
+  Yelp: yelpColDef,
   Default: defaultColDef,
 };
