+++ frontend/app/src/Components/Tables/CommentsTable/CommentsTableUrlConst.tsx
@@ -18,5 +18,7 @@ export enum CommentsTableUrlRequest {
   Instagram = '/instagram/comments',
   Twitter = '/twitter/comments',
   YouTube = '/youtube/comments',
-  Reddit = '/reddit/comments'
+  Reddit = '/reddit/comments',
+  GoogleReviews = '/googleReviews/comments',
+  Yelp = '/yelp/comments',
 }
