--- frontend/app/src/Components/Tables/ReviewsTable/ReviewsTableUrlConst.tsx
@@ -1,19 +1,0 @@
-interface GoogleReviewsReviewsTableResponse {
-  count: number;
-  data: {
-    id: number;
-    userName: string;
-    message: string;
-    likes: number;
-    sentimentAnalysis: string;
-    subjectivityAnalysis: string;
-    topicClassification: string;
-  }[];
-}
-
-export type ReviewsTableResponse = GoogleReviewsReviewsTableResponse;
-
-export enum ReviewsTableUrlRequest {
-  GoogleReviews = '/googleReviews/comments',
-  Yelp = '/yelp/comments',
-}
