+++ frontend/app/src/pages/ReviewApps/Yelp/YelpTab.tsx
@@ -6,7 +6,7 @@ import { AppNames } from '../../../Redux/Slices/webApp/webAppConstants';
 import SentimentPieChartWrapper from '../../../Components/Charts/PieChart/SentimentPieChartWrapper';
 import SubjectivityPieChartWrapper from '../../../Components/Charts/PieChart/SubjectivityPieChartWrapper';
 import CommentsWordCloud from '../../../Components/Charts/WordCloud/CommentsWordCloud';
-import ReviewsTable from '../../../Components/Tables/ReviewsTable/ReviewsTable';
+import CommentsTable from '../../../Components/Tables/CommentsTable/CommentsTable';
 
 interface Props {}
 
@@ -42,7 +42,7 @@ const YelpTab: React.FC<Props> = () => {
 
       <Grid item xs={12}>
         <CardCharts name={'Reviews'}>
-          <ReviewsTable appName={AppNames.Yelp} />
+          <CommentsTable appName={AppNames.Yelp} />
         </CardCharts>
       </Grid>
     </Grid>
